implemented graphql service module & fixed tslint errors
