+++ src/app/modules/core/graphql/interfaces/field-object.interface.ts
@@ -1,0 +1,6 @@
+/**
+ * GraphQL field object
+ */
+export interface IFieldObject {
+  [key: string]: string | string[] | IFieldObject;
+}
