+++ src/app/modules/core/graphql/interfaces/field.interface.ts
@@ -1,0 +1,6 @@
+import { IFieldObject } from './field-object.interface';
+
+/**
+ * GraphQL fields
+ */
+export type IFields = string | string[] | IFieldObject;
